requirements.txt

pymupdf
python-dotenv